tidy up gas processing
